64bit fix: VM_Call return value is a pointer

+++ code/server/sv_client.c
@@ -239,7 +239,7 @@ void SV_DirectConnect( netadr_t from ) {
 	int			challenge;
 	char		*password;
 	int			startIndex;
-	int		denied;
+	intptr_t		denied;
 	int			count;
 
 	Com_DPrintf ("SVC_DirectConnect ()\n");
